# implement GPU performance monitoring with backend-agnostic hooks