--- a/khora_engine_core/src/core/monitoring.rs
+++ b/khora_engine_core/src/core/monitoring.rs
@@ -20,6 +20,7 @@ use std::fmt::Debug;
 pub enum MonitoredResourceType {
     Vram,
     SystemRam,
+    GpuPerformance,
 }
 
 /// Represents the current usage and limit of a monitored resource.
@@ -30,6 +31,63 @@ pub struct ResourceUsageReport {
     pub total_capacity_bytes: Option<u64>,
 }
 
+/// Represents GPU performance metrics (times in microseconds for higher precision).
+#[derive(Debug, Clone, Copy, Default)]
+pub struct GpuPerformanceReport {
+    /// Frame number for tracking
+    pub frame_number: u64,
+    /// Individual hook timings in microseconds (indexed by GpuPerfHook as usize)
+    /// [FrameStart, MainPassBegin, MainPassEnd, FrameEnd]
+    pub hook_timings_us: [Option<u32>; 4],
+    /// CPU preparation time in microseconds
+    pub cpu_preparation_time_us: Option<u32>,
+    /// CPU submission time in microseconds
+    pub cpu_submission_time_us: Option<u32>,
+}
+
+impl GpuPerformanceReport {
+    /// Get timing for a specific GPU performance hook
+    pub fn get_hook_timing_us(
+        &self,
+        hook: crate::subsystems::renderer::api::common_types::GpuPerfHook,
+    ) -> Option<u32> {
+        self.hook_timings_us[hook as usize]
+    }
+
+    /// Get main pass duration (MainPassEnd - MainPassBegin)
+    pub fn main_pass_duration_us(&self) -> Option<u32> {
+        use crate::subsystems::renderer::api::common_types::GpuPerfHook;
+        match (
+            self.get_hook_timing_us(GpuPerfHook::MainPassBegin),
+            self.get_hook_timing_us(GpuPerfHook::MainPassEnd),
+        ) {
+            (Some(begin), Some(end)) if end >= begin => Some(end - begin),
+            _ => None,
+        }
+    }
+
+    /// Get total frame duration (FrameEnd - FrameStart)  
+    pub fn frame_total_duration_us(&self) -> Option<u32> {
+        use crate::subsystems::renderer::api::common_types::GpuPerfHook;
+        match (
+            self.get_hook_timing_us(GpuPerfHook::FrameStart),
+            self.get_hook_timing_us(GpuPerfHook::FrameEnd),
+        ) {
+            (Some(start), Some(end)) if end >= start => Some(end - start),
+            _ => None,
+        }
+    }
+
+    /// Set timing for a specific hook
+    pub fn set_hook_timing_us(
+        &mut self,
+        hook: crate::subsystems::renderer::api::common_types::GpuPerfHook,
+        timing_us: Option<u32>,
+    ) {
+        self.hook_timings_us[hook as usize] = timing_us;
+    }
+}
+
 /// The `ResourceMonitor` trait provides methods to monitor resources in a system.
 /// It includes methods to get the ID of the monitor, the type of resource being monitored,
 /// and a report of the current usage and limit of the resource.
@@ -37,4 +95,9 @@ pub trait ResourceMonitor: Send + Sync + Debug + 'static {
     fn monitor_id(&self) -> Cow<'static, str>;
     fn resource_type(&self) -> MonitoredResourceType;
     fn get_usage_report(&self) -> ResourceUsageReport;
+
+    /// Optional: Get GPU performance report for performance monitors
+    fn get_gpu_performance_report(&self) -> Option<GpuPerformanceReport> {
+        None
+    }
 }