--- a/khora_engine_core/src/lib.rs
+++ b/khora_engine_core/src/lib.rs
@@ -12,6 +12,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! Khora Engine Core Crate
+//!
+//! Provides the foundational runtime of KhoraEngine (main loop, event bus,
+//! rendering abstraction, math types, memory tracking) structured with the
+//! future Symbiotic Adaptive Architecture (SAA) in mind (see
+//! `../docs/architecture_design.md`).
+//!
+//! Main modules:
+//! * [`core`]: runtime engine (loop, event dispatch, timers, basic stats)
+//! * [`event`]: event definitions & decoupled publish / subscribe bus
+//! * [`math`]: vectors, matrices, quaternions, basic numeric helpers
+//! * [`memory`]: heap allocation tracking allocator
+//! * [`subsystems`]: pluggable subsystems (renderer `RenderSystem`, input, etc.)
+//! * [`window`]: OS window layer via `winit`
+//!
+//! Immediate extension points:
+//! * Add a new subsystem: create a module under `subsystems` and integrate in
+//!   `Engine` (event publication, future update hook).
+//! * Add instrumentation: extend `RenderStats` (and later the Core Metrics System).
+//!
+//! Rendering & GPU instrumentation:
+//! See `../docs/rendering/gpu_performance_monitoring.md` for comprehensive coverage of
+//! backend-agnostic monitoring, WGPU timestamp pipeline, and adaptive resize strategies.
+//!
+//! Quick design guidelines:
+//! * Keep the main loop non-blocking (avoid `std::thread::sleep` inside per-frame path)
+//! * Avoid panics outside tests; prefer `Result` or event signaling
+//! * Comments should focus on the "why" when logic is non-trivial
+//!
+//! Public API surface is intentionally small (Engine, EngineEvent, input events) to
+//! keep downstream usage minimal and stable during early iterations.
+
 pub mod core;
 pub mod event;
 pub mod math;