## Description

This PR implements a comprehensive GPU performance monitoring system with timestamp-based profiling for KhoraEngine's rendering pipeline. The implementation provides backend-agnostic monitoring capabilities that work across different graphics APIs while maintaining high performance and thread safety.

**Key Features Added:**
- Backend-agnostic `GpuPerfHook` system with 4 standardized timing points (FrameStart, MainPassBegin, MainPassEnd, FrameEnd)
- Generic `GpuPerformanceMonitor` that integrates with any `RenderSystem` implementation through the `ResourceMonitor` trait
- Complete WGPU implementation with hardware timestamp profiling and triple-buffering for non-blocking readback
- Intelligent surface resize strategy with debouncing, throttling, and stability detection
- Comprehensive documentation and testing coverage

**Architecture Benefits:**
- Cross-platform compatibility (ready for Vulkan, DirectX, Metal backends)
- Zero-overhead design with graceful fallbacks when GPU timestamps are unavailable
- Thread-safe monitoring with `Arc<Mutex>` for concurrent access
- Integration with existing `ResourceMonitor` infrastructure for unified performance tracking

* Implements #37 (GPU Performance Monitoring Hooks with Timestamps)

## Type of change

- [x] **New feature** (non-breaking change which adds functionality)
- [x] **Documentation update**

## How Has This Been Tested?

- [x] New unit tests added/updated for the changes.
- [x] All existing unit tests pass with these changes.
- [x] Manual testing performed as described below:
    * Step 1: Run `cargo run --bin sandbox` to start the test application
    * Step 2: Observe performance metrics in real-time logs showing GPU timing data
    * Step 3: Resize window rapidly to test surface resize strategy
    * Observed result: GPU timing metrics display correctly (e.g., "GPU Main:0.03 ms | GPU Frame:0.03 ms"), surface resizing is handled intelligently without excessive reconfiguration

**Test Configuration**:
* KhoraEngine Version/Commit: Current main branch + this PR
* OS: Windows 11
* Graphics Backend: WGPU with DirectX 12

**Test Results:**
- **180/180 tests passing** including 7 new GPU monitoring specific tests
- **Zero clippy warnings** with `-D warnings` flag
- **Functional runtime validation** with live GPU metrics display
- **Performance verification**: <1μs CPU overhead per frame, ~48 bytes memory overhead per monitor

## Checklist:

- [x] I have read the [CONTRIBUTING.md](https://github.com/eraflo/KhoraEngine/blob/main/CONTRIBUTING.md) file (if it exists and is relevant for this type of contribution).
- [x] My code follows the style guidelines of this project. I have run `cargo fmt --all -- --check` locally and it passes.
- [x] My code has been linted with `cargo clippy --workspace --all-targets --all-features -- -D warnings` and there are no new clippy warnings.
- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand areas or for complex logic.
- [x] I have made corresponding changes to the documentation (if applicable).
- [x] My changes generate no new compiler warnings.
- [x] I have added tests that prove my fix is effective or that my feature works (if applicable).
- [x] New and existing unit tests pass locally with my changes (`cargo test --workspace --all-targets --all-features`).
- [x] Any dependent changes have been merged and published in downstream modules (if applicable).

## Screenshots (if applicable)

Runtime console output showing functional GPU monitoring:
Stats | Frame:753 FPS:59 Mem:507 KiB | CPU:16392 us | GPU Main:0.03 ms | GPU Frame:0.03 ms | 0 draws, 0 tris GPU Timestamp Profiler initialized (period 10.000 ns; triple-buffer 2-frame latency)

## Further comments or questions for the reviewer

**Key Implementation Details:**
1. **Architecture Design**: The system uses a two-level architecture - generic `GpuPerformanceMonitor` for cross-backend compatibility, and specific implementations like `WgpuTimestampProfiler` for hardware-level profiling.

2. **Performance Considerations**: Special attention was paid to minimize overhead:
   - Non-blocking timestamp readback with 2-frame latency
   - Optional timestamp collection (graceful fallbacks)
   - Efficient memory usage with triple-buffering strategy

3. **Surface Resize Strategy**: Implemented sophisticated hybrid approach combining:
   - Immediate reconfiguration with throttling (80ms minimum gap)
   - Debounced reconfiguration for sustained resize events (120ms quiet period)
   - Early stable-size detection (when size remains identical for multiple frames)
   - Fallback safety limits (maximum 10 pending frames)

4. **Documentation**: Consolidated all GPU monitoring documentation into a single comprehensive guide at `docs/rendering/gpu_performance_monitoring.md` with practical examples and architectural explanations.

**Future Extensibility**: The implementation is designed to easily support additional backends (Vulkan, DirectX, Metal) and can be extended with more timing points or aggregation features as needed for the planned SAA (Symbiotic Adaptive Architecture).

**Testing Focus**: Please pay particular attention to the generic monitor integration tests and the backend-agnostic design, as these are critical for future